GRW-608 Move language picker to /components and some small fixes
