+++ src/client/pages/ConnectPayment/index.tsx
@@ -6,7 +6,7 @@ import { useHistory } from 'react-router'
 import { TopBar } from 'components/TopBar'
 import { Page } from 'components/utils/Page'
 import { SessionTokenGuard } from 'containers/SessionTokenGuard'
-import { LanguagePicker } from 'pages/Embark/LanguagePicker'
+import { LanguagePicker } from 'components/LanguagePicker/LanguagePicker'
 import { useTextKeys } from 'utils/textKeys'
 import { getUtmParamsFromCookie, TrackAction } from 'utils/tracking/tracking'
 import { PhoneNumber } from 'components/PhoneNumber/PhoneNumber'
